changing id type to long

+++ src/main/java/com/buaya/security/service/impl/HandicapService.java
@@ -48,7 +48,7 @@ public class HandicapService {
 		return handicapRepository.save(handicap);
 	}
 	
-	public Handicap findByUserId(int userId) {
+	public Handicap findByUserId(long userId) {
 		
 		Optional<Handicap> optional = handicapRepository.findByUserId(userId);
